refactor: move channel optimization problems into a new channel_opt package
Mirror the state-side layout (state_metrics vs state_opt) on the channel side.
The SDP optimization problems channel_exclusion, channel_distinguishability, and
unitary_exclusion lived in channel_metrics alongside the actual metrics
(fidelity, diamond distance, CB norms, relative entropies); move them into a new
toqito/channel_opt package so that *_metrics holds distance/similarity measures
and *_opt holds optimization problems, consistently for both states and channels.

channel_opt uses eager re-exports (matching every other subpackage). All import
paths, docstring cross-references, and the two channel_props callers
(is_channel_antidistinguishable, is_unitary_antidistinguishable) are updated;
channel_opt is registered in toqito/__init__.py. No backward-compatibility
shims: 'from toqito.channel_metrics import channel_exclusion' now raises.

Verified: import toqito clean, ruff clean, and the moved tests + remaining
channel_metrics tests + caller tests pass (129 passed, 7 skipped).
