+++ toqito/channel_metrics/__init__.py
@@ -7,6 +7,3 @@ from toqito.channel_metrics.channel_measured_relative_entropy import channel_mea
 from toqito.channel_metrics.channel_relative_entropy import channel_relative_entropy
 from toqito.channel_metrics.completely_bounded_trace_norm import completely_bounded_trace_norm
 from toqito.channel_metrics.completely_bounded_spectral_norm import completely_bounded_spectral_norm
-from toqito.channel_metrics.channel_distinguishability import channel_distinguishability
-from toqito.channel_metrics.channel_exclusion import channel_exclusion
-from toqito.channel_metrics.unitary_exclusion import unitary_exclusion
